before Use Docs Cursor Full length

+++ tools/clear_empty_dir.py
@@ -1,8 +1,9 @@
 import os,shutil
-dir_name = "./save"
+dir_name = "../save"
 dirs = os.listdir(dir_name)
 for dir in dirs:
-    cpt = sum([len(files) for r, d, files in os.walk(f'{dir_name}/{dir}')])
-    if cpt <= 0:
-        shutil.rmtree(f'{dir_name}/{dir}')
-        print(f'Remove {dir} ({cpt}) files')
+    if 'save_' in dir:
+        cpt = sum([len(files) for r, d, files in os.walk(f'{dir_name}/{dir}')])
+        if cpt <= 0:
+            shutil.rmtree(f'{dir_name}/{dir}')
+            print(f'Remove {dir} ({cpt}) files')
